Added more tests and fixes a number of issues
- Fixed issues where connections to other machines would not be closed
down correctly after Leave() was called.
- Fixed LocalNode not being serializable.  Technically it does serialize
now, but it serializes into a "fake" state.  When ITransparent objects
are received, their Node property is automatically updated to point at
the local node (as this is the property used to determine what node an
object should use for remoting).
- Refactored tests so that a seperate assembly (Dx.Runtime.Tests.Data)
has the post-processor run over it.